pyproject.toml
[tool.poetry]
name = "spider_test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.4"
drf-spectacular = "^0.24.2"
djangorestframework-simplejwt = "^5.2.1"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.0"
pre-commit = "^2.20.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
